Fix SIGALRM bug on Unix
We need to remove the timer, otherwise we will get a SIGALRM
later, after the lock was acquired. If this signal interrupts
a system call, we might get undefined results. In particular,
if the main R REPL is in a read(), and this is interrupted by
SIGALRM, interactive R processes quit.